+++ frontend/src/components/Postcode/Postcode.js
@@ -2,55 +2,30 @@ import React, { useEffect, useState } from "react";
 import DaumPostcode from "react-daum-postcode";
 import "./postcode.css";
 
-const Postcode = () => {
-  const [address, setAddress] = useState(""); // 주소
-  const [addressDetail, setAddressDetail] = useState(""); // 상세주소
-
-  const [isOpenPost, setIsOpenPost] = useState(false);
-
-  const onChangeOpenPost = () => {
-    setIsOpenPost(!isOpenPost);
-  };
-
-  const onCompletePost = (data) => {
-    let fullAddr = data.address;
-    let extraAddr = "";
+const Postcode = (props) => {
+  const complete = (data) => {
+    let fullAddress = data.address;
+    let extraAddress = "";
 
     if (data.addressType === "R") {
       if (data.bname !== "") {
-        extraAddr += data.bname;
+        extraAddress += data.bname;
       }
       if (data.buildingName !== "") {
-        extraAddr +=
-          extraAddr !== "" ? `, ${data.buildingName}` : data.buildingName;
+        extraAddress +=
+          extraAddress !== "" ? `, ${data.buildingName}` : data.buildingName;
       }
-      fullAddr += extraAddr !== "" ? ` (${extraAddr})` : "";
+      fullAddress += extraAddress !== "" ? ` (${extraAddress})` : "";
     }
-
-    setAddress(data.zonecode);
-    setAddressDetail(fullAddr);
-    setIsOpenPost(false);
-  };
-
-  const postCodeStyle = {
-    display: "block",
-    position: "relative",
-    top: "0%",
-    width: "400px",
-    height: "400px",
-    padding: "7px",
+    console.log(data);
+    console.log(fullAddress);
+    console.log(data.zonecode);
   };
 
   return (
-    <DaumPostcode>
-      {isOpenPost ? (
-        <DaumPostcode
-          style={postCodeStyle}
-          autoClose
-          onComplete={onCompletePost}
-        />
-      ) : null}
-    </DaumPostcode>
+    <div>
+      <DaumPostcode className="postmodal" autoClose onComplete={complete} />
+    </div>
   );
 };
